+++ cmdeploy/src/cmdeploy/service/chatmail-metadata.service.f
@@ -2,9 +2,11 @@
 Description=Chatmail dict proxy for IMAP METADATA
 
 [Service]
-ExecStart={execpath} /run/dovecot/metadata.socket vmail /home/vmail/mail/{mail_domain}
+ExecStart={execpath} /run/chatmail-metadata/metadata.socket /home/vmail/mail/{mail_domain}
 Restart=always
 RestartSec=30
+User=vmail
+RuntimeDirectory=chatmail-metadata
 
 [Install]
 WantedBy=multi-user.target